requirements.txt
spacy
transformers
torch
pandas
pymupdf
python-docx
scikit-learn
tqdm